Leave component in route file, create a test folder that wont generate routes
